requireのパスめっちゃ複雑なんでtraitもcontrollerのファイルにまとめた
