--- register/api/controllers/prompt/SetColumns.php
@@ -1,75 +1,0 @@
-<?php
-
-trait SetPromptColumns {
-    protected $columns;
-
-    protected function setPromptColumns() {
-        $this->columns = [
-            [
-                'name' => 'command_id',
-                'init' => -1,
-                'type' => PDO::PARAM_INT,
-            ],
-            [
-                'name' => 'command_name',
-                'init' => '',
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'command_jp',
-                'init' => '',
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'genre_id',
-                'init' => -1,
-                'type' => PDO::PARAM_INT,
-            ],
-            [
-                'name' => 'nsfw',
-                'init' => 'A',
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'variation',
-                'init' => null,
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'detail',
-                'init' => null,
-                'type' => PDO::PARAM_STR,
-            ],
-        ];
-    }
-
-    protected function setGenreColumns() {
-        $this->columns = [
-            [
-                'name' => 'genre_id',
-                'init' => -1,
-                'type' => PDO::PARAM_INT,
-            ],
-            [
-                'name' => 'genre_slag',
-                'init' => '',
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'genre_jp',
-                'init' => '',
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'caption',
-                'init' => null,
-                'type' => PDO::PARAM_STR,
-            ],
-            [
-                'name' => 'nsfw',
-                'init' => 'A',
-                'type' => PDO::PARAM_STR,
-            ],
-        ];
-    }
-}
